+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - Warn if users aren't on at least version 101 of Java 8
 - Change authlib version to 1.5.21 to fix issues with needing Java 8
+- Don't allow empty password when launching Minecraft without saved password